chore: enterpriseplanbadge as component
