+++ frontend/src/component/layout/components/EnterprisePlanBadge/EnterprisePlanBadge.tsx
@@ -1,0 +1,15 @@
+import { styled, Tooltip } from '@mui/material';
+import { EnterpriseBadge } from 'component/common/EnterpriseBadge/EnterpriseBadge';
+
+const StyledBadgeContainer = styled('div')(({ theme }) => ({
+    paddingLeft: theme.spacing(2),
+    display: 'flex',
+}));
+
+export const EnterprisePlanBadge = () => (
+    <Tooltip title='This is an Enterprise feature'>
+        <StyledBadgeContainer>
+            <EnterpriseBadge />
+        </StyledBadgeContainer>
+    </Tooltip>
+);
